pyproject.toml
[project]
name = "gemini-file-search-demo"
version = "0.1.0"
description = "Using Gemini File Search Tool to add RAG to an agentic application"
authors = [
    {name = "Dazbo (Darren Lester)", email = "derailed.dash@gmail.com"},
]
dependencies = [
    "google-adk>=1.15.0,<2.0.0",
    "google-cloud-logging>=3.12.0,<4.0.0",
    "google-cloud-aiplatform[adk]>=1.118.0,<2.0.0",
    "google-genai",
    "fastapi~=0.115.8",
    "uvicorn~=0.34.0",
]
requires-python = ">=3.12"


[project.optional-dependencies]
jupyter = [
    "jupyter>=1.0.0,<2.0.0",
    "ipython"
]
lint = [
    "ruff",
    "mypy",
    "codespell",
    "types-pyyaml",
    "types-requests",
]

[tool.ruff]
line-length = 120
target-version = "py312"
extend-exclude = ["*.ipynb"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C",  # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP", # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901", "B006", "B008", "W291", "E302"]

[tool.ruff.lint.isort]
known-first-party = ["app", "frontend"]

[tool.mypy]
disallow_untyped_calls = false
disallow_untyped_defs = false
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]

exclude = [".venv", ".*\\.ipynb"]

[tool.codespell]
ignore-words-list = "rouge"
skip = "uv.lock,.venv,./frontend,**/*.ipynb"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["app","frontend"]

